+++ riemann-java-client/src/main/java/io/riemann/riemann/client/netty3/HashedWheelTimerFactory.java
@@ -1,4 +1,4 @@
-package io.riemann.riemann.client;
+package io.riemann.riemann.client.netty3;
 
 import org.jboss.netty.util.HashedWheelTimer;
